split code btwn model and view
