+++ bower.json
@@ -2,7 +2,7 @@
     "name": "adapt-hotgrid",
     "repository": "git://github.com/cgkineo/adapt-hotgrid",
     "framework": "^2.0.0",
-    "version": "2.1.3",
+    "version": "2.1.4",
     "homepage": "https://github.com/cgkineo/adapt-hotgrid",
     "issues": "https://github.com/cgkineo/adapt-hotgrid/issues/",
     "displayName": "Hotgrid",
