+++ js/hotgridModel.js
@@ -1,0 +1,9 @@
+define([
+    'core/js/models/componentModel',
+    'core/js/models/itemsComponentModel'
+], function(ComponentModel, ItemsComponentModel) {
+
+    var HotgridModel = ItemsComponentModel.extend({});
+
+    return HotgridModel;
+});
